Enhance build process to automatically create interfaces and add copyrights (#10)
* Enhance build process to automatically create interfaces and add copyrights

In commit d5378affa3b636d117fb4602c5c395178ea70850 I added interfaces to
the generated code to allow mocking of functionality by users of the
client. I did not realize that whenever a change was made the files were
all regenerated, making those changes hard to keep.

This commit enhances the build process to automatically create and use the
interfaces as part of make gen, as well as adding copyrights to all the
generated files.
___
bretmckee-branch: automatically-generate-interfaces

* Add GOPRIVATE to v1/Makefile

The recent changes to vendoring made this necessary.

* Add v1/pkg/client/generic-error.go to .gitignore

* Regenerate the files

This includes the files generated by 'make gen' using the enhanced build
tools.