Enhance v1 volume-attach.go (#11)
* Enhance v1 volume-attach.go

This commit adds the structure necessary to create volume attachements.
There is currently no Quake volume attach create REST support, but in
order to generate it this file is required, so this change is being made
by itself. Once the Quake side is implemented, subsequent commits to
this repository will add client support.
___
bretmckee-branch: enhance-volume-attach

* address llabordehpe PR comments

* Fix lint found typo

* address llabordehpe PR comments